refactor: move reverify_human instruction
