+++ comptoken/src/instructions/mod.rs
@@ -1,13 +1,7 @@
 use solana_program::{account_info::AccountInfo, pubkey::Pubkey};
 
-mod mint_unchecked;
-pub use mint_unchecked::*;
-
-mod submit_proof;
-pub use submit_proof::*;
-
-mod initialize;
-pub use initialize::*;
+mod collect;
+pub use collect::*;
 
 mod create_user_data_account;
 pub use create_user_data_account::*;
@@ -15,14 +9,26 @@ pub use create_user_data_account::*;
 mod daily_distribution;
 pub use daily_distribution::*;
 
+mod flag_stale_account;
+pub use flag_stale_account::*;
+
 mod get_valid_blockhashes;
 pub use get_valid_blockhashes::*;
 
-mod collect;
-pub use collect::*;
+mod initialize;
+pub use initialize::*;
+
+mod mint_unchecked;
+pub use mint_unchecked::*;
 
 mod resize_user_data_account;
 pub use resize_user_data_account::*;
+
+mod reverify_human;
+pub use reverify_human::*;
+
+mod submit_proof;
+pub use submit_proof::*;
 
 mod unverify_human;
 pub use unverify_human::*;
@@ -32,9 +38,6 @@ pub use unverify_human_2::*;
 
 mod verify_human;
 pub use verify_human::*;
-
-mod flag_stale_account;
-pub use flag_stale_account::*;
 
 trait InstructionData: Sized {
     fn from_instruction_data(instruction_data: &[u8]) -> Result<Self, solana_program::program_error::ProgramError>;
